+++ toucan-shopping-modules/product/toucan-shopping-product-entity-vo/src/main/java/com/toucan/shopping/modules/product/entity/ShopProductApproveSku.java
@@ -99,8 +99,8 @@ public class ShopProductApproveSku {
     @JSONField(serializeUsing = ToStringSerializer.class)
     private Long shopCategoryId; //店铺商品分类ID
 
-    private Double roughWeight; //毛重
+    private BigDecimal roughWeight; //毛重
 
-    private Double suttle; //净重
+    private BigDecimal suttle; //净重
 
 }
